Renaming Activity and Workflow packages to Activities and Workflows
