+++ src/Temporalio.ApiDoc/api/index.md
@@ -1,9 +1,13 @@
 
 # API Documentation
 
-Namespaces:
+Commonly used namespaces:
 
-* [Temporalio](/api/Temporalio.html)
-* [Temporalio.Client](/api/Temporalio.Client.html)
-* [Temporalio.Testing](/api/Temporalio.Testing.html)
-* [Temporalio.Worker](/api/Temporalio.Worker.html)
+* [Temporalio](/api/Temporalio.html) - General purpose types
+* [Temporalio.Activities](/api/Temporalio.Activities.html) - Types for developing activities
+* [Temporalio.Client](/api/Temporalio.Client.html) - Client for accessing Temporal
+* [Temporalio.Converters](/api/Temporalio.Converters.html) - Data converters including support for custom conversion
+* [Temporalio.Exceptions](/api/Temporalio.Exceptions.html) - Well-known exceptions for cleints, activities, and workflows
+* [Temporalio.Testing](/api/Temporalio.Testing.html) - Testing environments and utilities
+* [Temporalio.Worker](/api/Temporalio.Worker.html) - Worker for running workflows and/or activities
+* [Temporalio.Workflows](/api/Temporalio.Workflows.html) - Types for developing workflows
