+++ src/Temporalio/Worker/ActivityWorker.cs
@@ -10,7 +10,7 @@ using System.Threading.Tasks;
 using Google.Protobuf;
 using Google.Protobuf.WellKnownTypes;
 using Microsoft.Extensions.Logging;
-using Temporalio.Activity;
+using Temporalio.Activities;
 using Temporalio.Converters;
 using Temporalio.Exceptions;
 using Temporalio.Worker.Interceptors;
